rental and reservation implementation
