+++ backend/nokia-innovative-project/src/main/java/com/nokia/library/nokiainnovativeproject/DTOs/RentalDTO.java
@@ -1,0 +1,4 @@
+package com.nokia.library.nokiainnovativeproject.DTOs;
+
+public class RentalDTO {
+}
