+++ backend/nokia-innovative-project/src/main/java/com/nokia/library/nokiainnovativeproject/controllers/RentalController.java
@@ -1,0 +1,4 @@
+package com.nokia.library.nokiainnovativeproject.controllers;
+
+public class RentalController {
+}
