+++ backend/nokia-innovative-project/src/main/java/com/nokia/library/nokiainnovativeproject/entities/Reservation.java
@@ -1,47 +1,4 @@
 package com.nokia.library.nokiainnovativeproject.entities;
 
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
-import javax.validation.constraints.Future;
-import javax.validation.constraints.NotNull;
-import java.io.Serializable;
-import java.util.Date;
-
-@Entity
-@Data
-@EntityListeners(AuditingEntityListener.class)
-public class Reservation implements Serializable {
-
-	@Id
-	@Setter(AccessLevel.NONE)
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@CreationTimestamp
-	@Setter(AccessLevel.NONE)
-	private Date rentalDate;
-
-	@NotNull(message = "The return date should be defined")
-	@Future(message = "The return date should be future")
-	private Date returnDate;
-
-	@ManyToOne(cascade = {CascadeType.DETACH,
-			CascadeType.MERGE,
-			CascadeType.PERSIST,
-			CascadeType.REFRESH},
-			fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@ManyToOne(cascade = {CascadeType.DETACH,
-			CascadeType.MERGE,
-			CascadeType.PERSIST,
-			CascadeType.REFRESH},
-			fetch = FetchType.LAZY)
-	@JoinColumn(name = "book_catalog_number")
-	private Book book;
-
+public class Reservation {
 }
