+++ backend/nokia-innovative-project/src/main/java/com/nokia/library/nokiainnovativeproject/repositories/RentalRepository.java
@@ -1,9 +1,9 @@
 package com.nokia.library.nokiainnovativeproject.repositories;
 
-import com.nokia.library.nokiainnovativeproject.entities.Rental;
+import com.nokia.library.nokiainnovativeproject.entities.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RentalRepository extends JpaRepository<Rental, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 }
