+++ backend/nokia-innovative-project/src/main/java/com/nokia/library/nokiainnovativeproject/services/RentalService.java
@@ -1,0 +1,4 @@
+package com.nokia.library.nokiainnovativeproject.services;
+
+public class RentalService {
+}
